fixed build error using next docs  like moved non-react component from pages folder to other folders to fix the build issue
