+++ next.config.js
@@ -7,6 +7,10 @@ const nextConfig = {
   images: {
     domains: ['res.cloudinary.com'],
   },
+  eslint: {
+    ignoreDuringBuilds: true,
+  },
+  pageExtensions: ['mdx', 'md', 'jsx', 'js', 'tsx', 'ts'],
 }
 
 module.exports = nextConfig
