+++ src/components/question/data.ts
@@ -1,4 +1,4 @@
-import { QuestionType } from '@/pages/questions/question.interface'
+import { QuestionType } from './question.interface'
 
 interface QuestionDataType {
   items: QuestionType[]
