+++ src/components/question/question.interface.ts
