+++ src/store/slice/index.ts
@@ -1,4 +1,4 @@
-import { questionReducer } from '@/pages/questions/question.slice'
+import { questionReducer } from "@/components/question/question.slice";
 
 export const reducer: any = {
   question: questionReducer,
